pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbf-sync-refactor"
version = "2.0.0"
authors = [
  { name="Ryner Lute" }
]
description = "Утилита синхронизации DBF-файлов с БД (SQLite / PostgreSQL) с поддержкой временных рядов"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "dbfread>=2.0.7",
  "psycopg2-binary>=2.9.9",
  "tqdm>=4.66.0"
]

[project.scripts]
dbf-sync = "dbf_sync_refactor:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"